#noticket Separate Simple and Dependent VFB
